+++ packages/tsurlfilter/CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [3.3.3] - 2025-05-22
 
+### Changed
+
+- Updated [@adguard/agtree] to `v3.2.0`.
+
 ### Fixed
 
 - Internal code for source rule finding from rulesets.